fix(yacht): extract shared maxImmediateScore, freeze Sets, fix bonus double-credit
Address code review findings on #2042 (story A1):

- Extract maxImmediateScore to aiHelpers.ts so ai.ts and probTables.ts share
  one implementation — prevents scoring logic from diverging between modules.
- Fix bonus double-credit: add curUpperSubtotal < 63 guard so the +35 credit
  only fires when the bonus is not yet earned. Previously any open upper
  category would receive phantom +35 credit after upperSubtotal reached 63.
- Freeze openCategories and filledCategories Sets in buildYachtInfoSet to
  match the dice-array freeze pattern and signal immutability intent.

All 239 yacht tests pass.
