+++ frontend/src/game/yacht/aiInfoSet.ts
@@ -141,8 +141,8 @@ export function buildYachtInfoSet(state: GameState, opponentScore: number): Yach
     dice: Object.freeze([...state.dice]),
     rollsUsed: state.rolls_used,
     rollsRemaining: Math.max(0, MAX_ROLLS_PER_TURN - state.rolls_used),
-    openCategories: open,
-    filledCategories: filled,
+    openCategories: Object.freeze(open) as ReadonlySet<Category>,
+    filledCategories: Object.freeze(filled) as ReadonlySet<Category>,
     categoriesRemaining: open.size,
     upperSubtotal: upperSub,
     toBonus,
